Moved tests to test directory
