+++ ShazzamForClothes/tests/context.py
@@ -1,0 +1,7 @@
+import os
+import sys
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+import img_preparation
+
+tests_root_dir = os.path.dirname(os.path.abspath(__file__))
